requirements.txt
Flask==2.0.3
Flask-Authlib-Client==0.0.1